move comment deletion to http DELETE method within same servlet
