+++ .gitignore
@@ -1,2 +1,4 @@
 target/
 .idea
+
+pastebin.txt
